torrent support is nearly working. to test the torrent support download the torrent manually:   addArchive("de", "2010-07-27", "", "wikipedia_de_2010-07-27.torrent", QUrl("http://evopedia.info/dumps/wikipedia_de_2010-07-27.torrent"), ret);   addArchive("sl", "2010-08-12", "", "wikipedia_sl_2010-08-12.torrent", QUrl("http://evopedia.info/dumps/wikipedia_sl_2010-08-12.torrent"), ret);   addArchive("small", "2010-08-14", "", "wikipedia_small_2010-08-14.torrent", QUrl("http://evopedia.info/dumps/wikipedia_small_2010-08-14.torrent"
then copy it to /tmp/whatver/ and issue a refresh downloads in the dumpSettings dialog, then select (with a right click) the de archive and
hit 'start torrent download' and select the directory the torrent is in (it will creat a new directory _in_ this directory). if you want resume
select the same directory again and not the newly created subdirectory.

can't test more right now as the seeds are down